Extract inner-class to package-private class
